Cambio de directorio de componente mapa
